refactor(http-types): split config value modules
