+++ crates/dcc-mcp-http-types/src/config/server.rs
@@ -1,0 +1,105 @@
+use std::net::IpAddr;
+
+use serde::{Deserialize, Serialize};
+
+// ── ServerConfig ───────────────────────────────────────────────────────────
+
+/// Core server identity & transport configuration.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(default)]
+pub struct ServerConfig {
+    /// Port to listen on. Default: 8765.
+    pub port: u16,
+
+    /// IP address to bind. Default: 127.0.0.1 (localhost only, per MCP security spec).
+    pub host: IpAddr,
+
+    /// MCP endpoint path. Default: `/mcp`.
+    pub endpoint_path: String,
+
+    /// Server name reported in MCP `initialize` response.
+    pub server_name: String,
+
+    /// Server version reported in MCP `initialize` response.
+    pub server_version: String,
+
+    /// Maximum concurrent SSE sessions. Default: 100.
+    pub max_sessions: usize,
+
+    /// Request timeout in milliseconds. Default: 30_000.
+    pub request_timeout_ms: u64,
+
+    /// Whether to enable CORS for browser-based MCP clients. Default: false.
+    pub enable_cors: bool,
+
+    /// How listener tasks (per-instance MCP endpoint and the optional
+    /// gateway) are driven. See [`ServerSpawnMode`] for the tradeoffs.
+    ///
+    /// Default: [`ServerSpawnMode::Ambient`]. PyO3-embedded users should
+    /// set this to [`ServerSpawnMode::Dedicated`] (the Python bindings do
+    /// so automatically). Fixes issue #303.
+    pub spawn_mode: ServerSpawnMode,
+
+    /// Maximum time to wait when self-probing a freshly bound listener to
+    /// confirm it is actually accepting connections before reporting
+    /// success. Applied per attempt; up to 5 attempts are made. Set to 0
+    /// to disable self-probing (not recommended). Default: 200.
+    pub self_probe_timeout_ms: u64,
+}
+
+impl Default for ServerConfig {
+    fn default() -> Self {
+        Self {
+            port: 8765,
+            host: IpAddr::V4(std::net::Ipv4Addr::LOCALHOST),
+            endpoint_path: "/mcp".to_string(),
+            server_name: "dcc-mcp".to_string(),
+            server_version: env!("CARGO_PKG_VERSION").to_string(),
+            max_sessions: 100,
+            request_timeout_ms: 30_000,
+            enable_cors: false,
+            spawn_mode: ServerSpawnMode::Ambient,
+            self_probe_timeout_ms: 200,
+        }
+    }
+}
+
+// ── ServerSpawnMode ────────────────────────────────────────────────────────
+
+/// How the server and gateway HTTP listeners are driven.
+///
+/// Fixes **issue #303** — under PyO3-embedded interpreters (Maya on Windows),
+/// `tokio::spawn` onto a multi-threaded runtime that no longer has an active
+/// driver can cause background accept loops (specifically the gateway
+/// listener) to be starved of scheduling time. The per-instance listener
+/// survives because its accept loop is "warmed up" during the initial
+/// `block_on`, but the gateway listener — spawned via an extra `tokio::spawn`
+/// + `tokio::join!` layer — never gets its turn.
+///
+/// `ServerSpawnMode::Dedicated` avoids the failure mode entirely by running
+/// each HTTP listener on its own OS thread that owns a `current_thread`
+/// Tokio runtime. That thread is scheduled by the OS, not by a shared
+/// worker pool, and cannot be starved by a hanging block_on elsewhere.
+///
+/// | Mode | When to use | Behaviour |
+/// |------|-------------|-----------|
+/// | `Ambient`   | Standalone binary (`dcc-mcp-server`, library tests) | Spawns `axum::serve` onto the caller's Tokio runtime via `tokio::spawn`. |
+/// | `Dedicated` | Python bindings (`PyMcpHttpServer`) / embedded DCC hosts | Each listener gets its own OS thread + `current_thread` runtime. Immune to PyO3 worker starvation. |
+///
+/// Defaults: `Ambient`. The Python bindings override this to `Dedicated`
+/// automatically when constructing `McpHttpServer` via `PyMcpHttpServer`.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Default, Serialize, Deserialize)]
+#[serde(rename_all = "snake_case")]
+pub enum ServerSpawnMode {
+    /// Spawn listeners as background tasks on the caller's Tokio runtime.
+    /// Correct for `#[tokio::main]` binaries that keep a thread in the
+    /// runtime for the process lifetime.
+    #[default]
+    Ambient,
+
+    /// Spawn each listener on a dedicated OS thread with its own
+    /// `current_thread` runtime. Correct for PyO3-embedded interpreters
+    /// where the parent runtime's worker pool cannot be relied upon after
+    /// `block_on` returns.
+    Dedicated,
+}
